chore: more table logic to hook
